% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlUtils-package.R
\name{bw_silver}
\alias{bw_silver}
\title{Silverman default bandwidth calculation}
\usage{
bw_silver(xy, ess = NULL)
}
\arguments{
\item{xy}{Data coordinates}

\item{ess}{An effective sample size. If left as \code{NULL}, \code{nrow(xy)} is used.}
}
\description{
Silverman default bandwidth calculation
}
