% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_utils_funcs.R
\name{crw2ctmm}
\alias{crw2ctmm}
\title{Convert sf output from crawl to ctmm object}
\usage{
crw2ctmm(data, time_name, silent = TRUE)
}
\arguments{
\item{data}{\code{sf} data object}

\item{silent}{Locical. Should messages from \code{ctmm} be printed. Defaults to
\code{silent = TRUE}.}

\item{time.name}{Character. The name of the column of location times}
}
\description{
Takes \code{crawl::crwPredict} or \code{crawl::crwPostIS}
objects converted to \code{sf} objects by \code{crawl::crw_as_sf} to data
objects that can be used within the \code{ctmm} package.
}
\author{
Devin S. Johnson
}
