% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_fit_pred_sample_funcs.R
\name{cu_add_argos_cols}
\alias{cu_add_argos_cols}
\title{Add columns for modeling FastGPS and ARGOS error structure in the same
telemetry deployment}
\usage{
cu_add_argos_cols(x)
}
\arguments{
\item{x}{Data frame containing location telemetry data and Argos quality
information. See 'Details' for a description of the necessary data column names.}
}
\description{
Columns are added to the telemetry data set so that multiple data types
can be used simultaneously within the same animal: FastGPS, Argos least-squares, and
Argos Kalman filter.
}
\details{
To use this function the data set must contain the following columns with
exact names: (1) \code{"type"}, indicate the type of location,
(2) \code{"quality"} which indicates the Argos quality of the location, (3) The
Argos KF diagnostics: \code{"error_semi_major_axis"}, \code{"error_semi_minor_axis"}, and
\code{"error_ellipse_orientation"}. If there are no Argos KF locations, the associated columns
need not be present. Values of \code{type} are 'FastGPS', 'Argos', or 'known'. For 'FastGPS' an error radius
of 100m is assumed. For 'known' an error radius of 20m is assumed. Values of
\code{quality} must be '3','2','1','0','A',or 'B' for \code{type=='Argos'} locations and '4'-'10' for \code{type=='FastGPS'}.
For other types it is ignored.
}
\author{
Devin S. Johnson
}
