% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_add_migration.R
\name{cu_add_bouts}
\alias{cu_add_bouts}
\title{Add migration detection results to original location data}
\usage{
cu_add_bouts(data, bout_tbl)
}
\arguments{
\item{data}{Original data used by \code{cu_migration_det}.}

\item{bout_tbl}{Results table produced by \code{cu_migration_det}.}
}
\description{
Add bout data from \code{migr_evt}
See \code{\link[crawlUtils]{cu_bout_det}}
}
\author{
Devin S. Johnson
}
