% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_add_gaps.R
\name{cu_add_gaps}
\alias{cu_add_gaps}
\title{Label telemetry bout segments}
\usage{
cu_add_gaps(x, gap = 7, time_unit = "days", ...)
}
\arguments{
\item{x}{A data set of telemetry locations and times.}

\item{gap}{specified maximum time interval for which a new bout is started. Defaults
to \code{gap=7} days.}

\item{time_unit}{Unit of time of the gap specification. Defaults to \code{"day"}.}

\item{...}{Ignored arguments}
}
\description{
Label segments of telemetry data such that gaps between observed
locations are less than or equal to the interval specified.
}
\author{
Devin S. Johnson
}
