% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_add_migration.R
\name{cu_add_migration}
\alias{cu_add_migration}
\title{Add migration detection results to original location data}
\usage{
cu_add_migration(data, migr_tbl)
}
\arguments{
\item{data}{Original data used by \code{cu_migration_det}.}

\item{migr_tbl}{Results table produced by \code{cu_migration_det}.}
}
\description{
Add \code{migr_evt}
See \code{\link[crawlUtils]{cu_migration_det}}
}
\author{
Devin S. Johnson
}
