% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_as_telem.R
\name{cu_as_telem}
\alias{cu_as_telem}
\title{Convert Wildlife Computers data imported with
\verb{\link[crawlUtils]\{cu_read_wc_dirs\}} to a \code{telemetry} object from the
\code{ctmm} package.}
\usage{
cu_as_telem(locs_df, ...)
}
\arguments{
\item{locs_df}{A data frame output by the function \verb{\link[crawlUtils]\{cu_read_wc_dirs\}}}

\item{...}{Additional arguments to be passed to \verb{\link{ctmm}{as.temeletry}}}
}
\description{
Convert Wildlife Computers data imported with
\verb{\link[crawlUtils]\{cu_read_wc_dirs\}} to a \code{telemetry} object from the
\code{ctmm} package.
}
\author{
Josh M. London, Devin S. Johnson
}
