% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_bout_det_mbc.R
\name{cu_bout_det_mbc}
\alias{cu_bout_det_mbc}
\title{Movement bout detection}
\usage{
cu_bout_det_mbc(
  data,
  min_disp,
  migr_speed_cut = 1,
  min_bout_len = 3,
  Gmax = 3,
  grid_res = "day",
  base = "first"
)
}
\arguments{
\item{data}{An \code{sf} data set of locations with times noted by the 'datetime' column}

\item{min_disp}{The minimum dispersion rate to be considered a migration interval,
e.g. 10 for a 10km dispersion minimum.}

\item{migr_speed_cut}{The minimum speed to categorize a movement bout as "migratory" Defaults to 1 km/h.}

\item{min_bout_len}{Minimum length for each bout, defaults to 3}

\item{Gmax}{Maximum number of movement clusters, defaults to 3.}

\item{grid_res}{The temporal resolution at which migrations are detected. e.g., "day" (default) implies migration
start and end is detected on a daily resolution.}

\item{base}{The location at which dispersion is measured. Can be one of \code{"first"} (first location),
\code{"last"} (final location), or some other \code{sf::sfc} point location.}
}
\description{
Creates a data table that indicates the times of different
bouts of movement. This method uses changes in the overall dispersion rate of the animal
from the 'base' time to detect changes in overall movement from small scale local movement
to large scale migration.
}
\author{
Devin S. Johnson
}
