% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_bout_summary.R
\name{cu_bout_summary}
\alias{cu_bout_summary}
\title{Create a Table of Start and End Times for each Bout in a telemetry data set}
\usage{
cu_bout_summary(x, ...)
}
\arguments{
\item{x}{A telemetry data frame containing a \code{bout_id} column.}

\item{...}{Ignored arguments}
}
\description{
Summarizes a telemetry dataframe with a bout column into
a data frame with start and end times for each bout. For use with
\code{\link[crawlUtils:cu_join_interval_tbl]{cu_join_interval_tbl}}.
}
\author{
Devin S. Johnson
}
