% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel_funcs.R
\name{cu_crw2ctmm}
\alias{cu_crw2ctmm}
\title{Convert sf output from crawl to ctmm object}
\usage{
cu_crw2ctmm(data, silent = FALSE)
}
\arguments{
\item{data}{\code{sf} data object from \code{\link[crawl:crw_as_sf]{crawl::crw_as_sf}}}

\item{silent}{Logical. Should messages from \code{ctmm} be printed. Defaults to
\code{silent = FALSE}.}
}
\description{
Takes \code{crawl::crwPredict} or \code{crawl::crwPostIS}
objects converted to \code{sf} objects by \code{\link[crawl:crw_as_sf]{crawl::crw_as_sf}} and converts them to data
objects that can be used within the \code{ctmm} package.
}
\author{
Devin S. Johnson
}
