% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_fit_pred_sample_funcs.R
\name{cu_crw_argos}
\alias{cu_crw_argos}
\title{Batch Fitting CTCRW Models for Argos (and FastGPS) Data}
\usage{
cu_crw_argos(
  data_list,
  move_phase = NULL,
  bm = FALSE,
  crw_control = NULL,
  fixPar = NULL,
  ...
)
}
\arguments{
\item{data_list}{A list of data sets. Will also accept a single \code{sf} data frame as well.}

\item{move_phase}{An optional character value indicating a factor variable column in the data that
designates different movement phases.}

\item{bm}{Fit a Brownian Motion model rather than in integrated OU model. Defaults to \code{bm = FALSE}.}

\item{crw_control}{A named list passed to \code{\link[crawl]{crwMLE}} for optimization.
There is one additional parameter \code{lambda}. If added \code{crw_control$lambda}
is rate parameter used for the soft equality constraint on class \code{0}, \code{A},
 and \code{B} parameters for Argos LS observations in a mixed KF/LS dataset.}

\item{fixPar}{An alternative to the default set of fixed parameter values. Care should be taken
when substituting different values. Make sure you know what you're doing because it can be easily
broken}

\item{...}{Additional arguments passed to the \code{\link[foreach]{foreach}} function, e.g.,
for error handling in the loop.}
}
\description{
A basic CTCRW model is fitted to a list of data sets where each
element in the list represents the data for a single individual or deployment.
}
