% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_crw_covmat.R
\name{cu_crw_covfun}
\alias{cu_crw_covfun}
\title{Create covariance function for a fitted CRW model object}
\usage{
cu_crw_covfun(x)
}
\arguments{
\item{x}{A crwFit object created by a call to \code{\link[crawl]{crwMLE}}}
}
\description{
A function is created to evaluate the covariance function of the fitted
CRW movement model
}
\details{
The function returns a function to evaluate the covariance of the fitted Integrated
Ornstein-Ulenbeck movement model. The returned function has 3 arguments: (1) \code{t1}
and (2) \code{t2} both vectors of times to evaluate the covariance function of the fitted
IOU model, and (3) \code{E} (defaults to \code{E=0}). Which is the "zero" time of the process.
Typically \code{E} will be the time of the first observation.
}
\references{
Taylor, J. M., Cumberland, W. G., & Sy, J. P. (1994).
A stochastic model for analysis of longitudinal AIDS data. Journal of the
American Statistical Association, 89(427), 727-736.

Johnson, D. S., London, J. M., Lea, M. A., & Durban, J. W. (2008).
Continuous‐time correlated random walk model for animal telemetry data.
Ecology, 89(5), 1208-1215.
}
\author{
Devin S. Johnson
}
