% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_funcs.R
\name{cu_crw_covmat}
\alias{cu_crw_covmat}
\title{Calculate correlation matrix for a set of times from a CRW covariance function}
\usage{
cu_crw_covmat(x, corr = TRUE, cf, E = 0)
}
\arguments{
\item{x}{Either a \code{crwFit} object from a call to \code{\link[crawl]{crwMLE}}, or a
vector of times.}

\item{corr}{Should the function return a correlation or covariance matrix? Defaults
to \code{corr = TRUE}.}

\item{cf}{Covariance function created from a \code{crwFit} object with a call to
\code{\link{cu_crw_covfun}}.}

\item{E}{The 'zero' time used for the covariance function. Defaults to \code{E = 0}.
(See \code{\link{cu_crw_covfun}}).}
}
\description{
Using a correlation function created by \code{\link{cu_crw_covfun}}
from a fitted CRW model a covariance (correlation) matrix is created for observations
at the user provided times.
}
\details{
If \code{x} is a \code{crwFit} object, then the \code{cf} and \code{E}
arguments are ignored. The resulting matrix is the covariance (correlation) matrix
for the observed location times conditioned on the first location. Therefore,
for n observed locations an (n-1) by (n-1) matrix will result. This is most
useful for the effective sample size computation for a kernel density estimate.
}
\author{
Devin S. Johnson
}
