% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_funcs.R
\name{cu_crw_ess}
\alias{cu_crw_ess}
\title{Calculate Effective Sample Size for a Set of CRW locations}
\usage{
cu_crw_ess(fit, aug = NULL)
}
\arguments{
\item{fit}{A \code{crwFit} object (See \code{\link[crawl]{crwMLE}}).}

\item{aug}{Either a \code{\link[crawl]{crwPredict}} or \code{\link[crawl]{crwPostIS}} objects
from which the extra \code{predTime} location times will be used in the calculation.
The \code{\link[crawl]{crw_as_sf}} transformed versions of these objects will also work.}
}
\description{
Estimates the number of independent locations in a CRW data set
using the mutual information method of Bartoszek (2016).
}
\details{
Uses the "mutual information" formulation of Bartoszek (2016) to calculate the
equivalent number of independent animal locations.
}
\references{
Bartoszek, K. (2016). Phylogenetic effective sample size.
Journal of Theoretical Biology. 407:371-386. (See https://arxiv.org/pdf/1507.07113.pdf).
}
\author{
Devin S. Johnson
}
