% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_fit_pred_sample_funcs.R
\name{cu_crw_sample}
\alias{cu_crw_sample}
\title{Batch CRW Posterior Path Simulation For Multiple Animals}
\usage{
cu_crw_sample(size = 8, fit_list, predTime, barrier = NULL, vis_graph = NULL)
}
\arguments{
\item{size}{The number of posterior draws. Defaults to 8 (See Details).}

\item{fit_list}{A list of CRW fit objects}

\item{predTime}{A character string describing the desired frequency of prediction,
e.g., \code{predTime="1 hour"} or \code{predTime="15 min"}.}

\item{barrier}{An \code{sf} polygon object representing areas where the animal cannot access.}

\item{vis_graph}{A visibility graph constructed with the R package \code{pathroutr}, which is used
to reroute paths around barriers.}
}
\description{
Uses a list of CRW fitted models and desired simulation times
to make draws from the location (and velocity) posterior distribution for telemetered animals.
}
\details{
The R package \code{pathroutr} is necessary for use of the \code{barrier} rerouting.
it can be installed with the command
\code{install.packages('pathroutr', repos='https://jmlondon.r-universe.dev')}.
See 'https://github.com/jmlondon/pathroutr' for a description of use and constructing the
viability \code{vis_graph}.
}
\author{
Devin S. Johnson
}
