% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_disp_rate.R
\name{cu_disp_rate}
\alias{cu_disp_rate}
\title{Movement bout detection}
\usage{
cu_disp_rate(bout_tbl, data, grid_res = "day", base = "first")
}
\arguments{
\item{bout_tbl}{A bout table for an individual animal created by \code{\link[crawlUtils]{cu_bout_det}}}

\item{data}{An \code{sf} data set of locations with times noted by the 'datetime' column}

\item{grid_res}{The temporal resolution at which migrations are detected. e.g., "day" (default) implies migration
start and end is detected on a daily resolution.}

\item{base}{The location at which dispersion is measured. Can be one of \code{"first"} (first location),
\code{"last"} (final location), or some other \code{sf::sfc} point location.}
}
\description{
Creates a data table that indicates the times of different
bouts of movement. This method uses changes in the overall dispersion rate of the animal
from the 'base' time to detect changes in overall movement from small scale local movement
to large scale migration.
}
\author{
Devin S. Johnson
}
