% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_empirical_ud.R
\name{cu_empirical_ud}
\alias{cu_empirical_ud}
\title{Calculate an empirical utilization distribution}
\usage{
cu_empirical_ud(grid, pts, average = TRUE)
}
\arguments{
\item{grid}{a grid created from \code{\link[=cu_ud_grid]{cu_ud_grid()}}}

\item{pts}{an sf point object}

\item{average}{should an average UD be created from 1 or more pts layers}
}
\value{
grid with an additional column, mean_pts
}
\description{
Calculate an empirical utilization distribution
}
