% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_extract_obs.R
\name{cu_extract_obs}
\alias{cu_extract_obs}
\title{Extract predicted locatiins closest in time to observed locations.}
\usage{
cu_extract_obs(x, obs)
}
\arguments{
\item{x}{A \code{crwPredict} or \code{crwIS} object. Time column must be named \code{datetime}.}

\item{obs}{Observed locations. Time column must be named \code{datetime}.}
}
\description{
This function is intended for use after a predicted track or posterior simulation
is routed around barriers using the \code{{pathroutr}} package.
}
\author{
Devin S. Johnson
}
