% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_mbc_ud.R
\name{cu_gmm_ud}
\alias{cu_gmm_ud}
\title{Gaussian Mixture Model Utilization Density Estimation}
\usage{
cu_gmm_ud(pts, grid, ess = nrow(pts), type = "original", mclust_args = list())
}
\arguments{
\item{pts}{A \code{\link[crawl]{crwPredict}} or \code{\link[crawl]{crwPostIS}} object, or
their 'sf' versions (See \link[crawl]{crw_as_sf}).}

\item{grid}{An \code{\link[sf]{sf}} data frame containing the desired grid location for UD estimation.}

\item{ess}{Effective sample size.}

\item{type}{Form of the return type. Can be \code{"original"} to have the UD returned in the same form as the \code{grid}
argument. Set \code{type="vector"} to return only a vector of UD values. Finally, set \code{type="gmm"}
to return the fitted \code{\link[mclust]{densityMclust}} object.}

\item{mclust_args}{A named list of additional arguments to pass to \code{\link[mclust]{densityMclust}}.}
}
\description{
Calculates a Gaussian KDE for \code{crawl} simulations and predictions.
}
\author{
Devin S. Johnson
}
