% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_gmm_ud_sample.R
\name{cu_gmm_ud_sample}
\alias{cu_gmm_ud_sample}
\title{Kernel density estimation for a posterior sample list}
\usage{
cu_gmm_ud_sample(
  smp_list,
  grid,
  ess = NULL,
  return_gmm = FALSE,
  mclust_args = list()
)
}
\arguments{
\item{smp_list}{A list of posterior track samples from \code{\link[crawl]{crwPostIS}} after
transformation via \code{\link[crawl]{crw_as_sf}}.}

\item{grid}{An \code{\link[sf]{sf}} data frame containing the desired grid location for UD estimation.}

\item{ess}{Effective sample size.}

\item{return_gmm}{Logical. Return a list of Gaussian mixture model fits. Default is \code{FALSE}.}

\item{mclust_args}{Named list of additional arguments to pass to \code{\link[mclust]{densityMclust}}}
}
\description{
Kernel density estimation for a posterior sample list
}
\author{
Devin S. Johnson
}
