% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_GPSeq_clus.R
\name{cu_GPSeq_clus}
\alias{cu_GPSeq_clus}
\title{Add cluster information from the {GPSeq_clus} package.}
\usage{
cu_GPSeq_clus(
  x,
  search_radius_m,
  window_days,
  clus_min_locs = 2,
  centroid_calc = "mean",
  show_plots = c(FALSE, "mean"),
  scale_plot_clus = TRUE,
  store_plots = FALSE,
  season_breaks_jul = NA,
  daylight_hrs = NA
)
}
\arguments{
\item{x}{Output from \code{cu_crw_predict} or \code{cu_crw_sample}.}

\item{search_radius_m}{Search radius (meters) from cluster centroid when building clusters.}

\item{window_days}{Temporal window (days) to search for new locations from the most recent location in a cluster}

\item{clus_min_locs}{Minimum number of locations required to form a cluster. Default is 2.}

\item{centroid_calc}{Method for recalculating centroids when actively building clusters - e.g., "median" or "mean" (default). Not to be confused with
plotting the "mean" or "median" centroid once a cluster has been built.}

\item{show_plots}{Vector of TRUE/FALSE for plotting followed by plotting argument for the "median" or "mean" centroid - e.g., c(TRUE, "mean") (default)}

\item{scale_plot_clus}{When plotting, scale cluster markers based on number of locations (TRUE/FALSE).}

\item{store_plots}{When plotting, also assign map outputs to global environment (TRUE/FALSE).}

\item{season_breaks_jul}{Ascending numeric vector of julian days (0-365) used to classify by season/parturition/hunting seasons etc.
e.g., c(121, 274, 305) result may be: 1 Nov - 30 Apr (winter = 0), 1 May - 31 Aug (summer = 1), 1 Oct - 31 Oct (hunting season = 2)}

\item{daylight_hrs}{Manually set start and stop hours (0-24) to classify day and night locations. - e.g. c(6,18) would classify 6AM - 6PM as daylight hrs.
NA (default) uses 'suncalc' package to convert cluster location and time to be classified based on specific specific sunrise and sunset times.}
}
\description{
The function is simply a wrapper for the \code{\link[GPSeqClus]{GPSeq_clus}} function. All of the
arguments are the same except for the \code{dat} argument (here \code{x}). The \code{x} argument
must be an object output from \code{\link[crawlUtils]{cu_crw_predict}} or \code{\link[crawlUtils]{cu_crw_sample}} with
\code{as_sf=TRUE}. In addition, here \code{show_plots=c(FALSE, "mean")} as the default. As the function executes
there are a number of progress bars, they can be ignored.
}
\author{
Devin S. Johnson
}
