% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_hud.R
\name{cu_hud}
\alias{cu_hud}
\title{Highest utilization density}
\usage{
cu_hud(ud, alpha = 0.9)
}
\arguments{
\item{ud}{A \code{ud_df} object output from \code{\link{cu_kde_ud}}.}

\item{alpha}{The percent cutoff for the highest utilization probability cells. Defaults to \code{alpha = 0.9}.}
}
\description{
The lowest 1-alpha percent of a utilization distribution is removed
to give a highest alpha\\% UD which can be used as a home range estimate, or just
reduce spatial extent of UDs for a animal with a small spatial scale of use relative to the study area
}
\author{
Devin S. Johnson
}
