% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_funcs.R
\name{cu_kde_ud_sample}
\alias{cu_kde_ud_sample}
\title{Kernel density estimation for a posterior sample list}
\usage{
cu_kde_ud_sample(
  smp_list,
  grid,
  kern = "iso",
  ess = NULL,
  norm = TRUE,
  B = NULL
)
}
\arguments{
\item{smp_list}{A list of posterior track samples from \code{\link[crawl]{crwPostIS}} after
transformation via \code{\link[crawl]{crw_as_sf}}.}

\item{grid}{An \code{\link[sf]{sf}} data frame containing the desired grid location for UD estimation.}

\item{kern}{Type of covariance matrix for the Gaussian kernels.}

\item{ess}{Effective sample size.}

\item{norm}{Logical. Should each individual kernel be normalized to
sum-to-one over the locations in \code{grid}. Defaults to \code{kern = TRUE}}

\item{B}{Kernel covariance matrix. Defaults to \code{B = NULL} and a effective
sample size calculation is used for a plugin 2d Gaussian kernel.}
}
\description{
Kernel density estimation for a posterior sample list
}
\author{
Devin S. Johnson
}
