% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_migration_det.R
\name{cu_migration_det}
\alias{cu_migration_det}
\title{Migration detection}
\usage{
cu_migration_det(
  data,
  min_disp,
  migr_disp_cut = 1,
  min_bout_len = 3,
  grid_res = "day",
  base = "first",
  max_k = 100
)
}
\arguments{
\item{data}{An \code{sf} data set of locations with times noted by the 'datetime' column}

\item{min_disp}{The minimum dispersion rate to be considered a migration interval,
e.g. 10 for a 10km dispersion minimum.}

\item{migr_disp_cut}{The minimum per day dispersal rate to categorize a movement bout as "migratory" Defaults to 1.}

\item{min_bout_len}{The mimimum length of time that a migration or non-migration event
will take, e.g., 7 implies a minimum of 7 time intervals for a bout.}

\item{grid_res}{The temporal resolution at which migrations are detected. e.g., "day" (default) implies migration
start and end is detected on a daily resolution.}

\item{base}{The location at which dispersion is measured. Can be one of \code{"first"} (first location),
\code{"last"} (final location), or some other \code{sf::sfc} point location.}

\item{max_k}{The maximum degrees of freedom used by \code{mgcv::gam} to model dispersion and estimate
the derivative of the dispersion function.}
}
\description{
Creates a data table that indicates the times of different
bouts of movement. This method uses changes in the overall dispersion rate of the animal
from the 'base' time to detect changes in overall movement from small scale local movement
to large scale migration.
}
\author{
Devin S. Johnson
}
