% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_sf_kde.R
\name{cu_sf_kde}
\alias{cu_sf_kde}
\title{Spatial kernel density estimate}
\usage{
cu_sf_kde(x, w = NULL, bw = NULL, ref, ess = NULL, mask = FALSE)
}
\arguments{
\item{x}{sf POINT object}

\item{w}{Optional values, associated with \code{x} coordinates,
to be used as weights}

\item{bw}{Standard deviation scale bandwidth of Gaussian Kernel, must be units
of \code{x} projection.}

\item{ref}{A terra SpatRaster}

\item{ess}{A effective sample size to use instead of \code{nrow(x)} for determining the default bandwidth.}

\item{mask}{(TRUE/FALSE) mask resulting raster if ref is provided
as a SpatRaster}
}
\value{
a terra SpatRaster class object containing kernel density estimate
}
\description{
A weighted or unweighted Gaussian Kernel Density estimate
for point spatial data
}
\details{
Please note that ks methods for estimation has been reverted to the Gussian method proposed
in Venables & Ripley (2002). There was not enought evendence that the Chacon & Duong (2018)
multivariate method(s) for bandwidth selection and kernal estimation were suitable for
spatial random fields.
}
\references{
Duong, T. & Hazelton, M.L. (2005) Cross-validation bandwidth matrices for multivariate
kernel density estimation. Scandinavian Journal of Statistics, 32, 485-506.

Wand, M.P. & Jones, M.C. (1994) Multivariate plug-in bandwidth selection. Computational
Statistics, 9, 97-116.

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S.
Fourth edition. Springer.
}
\author{
Jeffrey S. Evans  \href{mailto:jeffrey_evans@tnc.org}{jeffrey_evans@tnc.org} and Devin S. Johnson \href{mailto:devin.johnson@noaa.gov}{devin.johnson@noaa.gov}
}
