% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_ud_grid.R
\name{cu_ud_grid}
\alias{cu_ud_grid}
\title{Create a grid for kernel density estimation of animal UDs}
\usage{
cu_ud_grid(bb, barrier = NULL, remove_holes = TRUE, clique = FALSE, ...)
}
\arguments{
\item{bb}{An \code{sf} bounding box or something that can be coerced into a
bounding box via \code{\link[sf]{st_bbox}} function.}

\item{barrier}{\code{sf} or \code{sfc} polygon data which represents areas for
which the animal cannot travel. Use will only be calculated on the outside of
each polygon in \code{barrier}.}

\item{remove_holes}{Remove holes in grid. Can sometimes speed up computations if small islands or lakes are removed.
defaults to \code{TRUE}.}

\item{clique}{Logical. Should separated segments be noted so they might be removed later. Only applicable
when a barrier is provided.}

\item{...}{Additional arguments passed to \code{\link[sf]{st_make_grid}} which
is used to construct the base grid.}
}
\description{
A \code{sf} data frame is created with polygon grid cells to evaluate a Gaussian
KDE at the centroids
}
\author{
Devin S. Johnson
}
