% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_vel_B.R
\name{cu_vel_B}
\alias{cu_vel_B}
\title{Compute velocity based kernel density estimate bandwidth}
\usage{
cu_vel_B(
  pts,
  ess = list(nrow(pts), rep(1/nrow(pts), nrow(pts))),
  vel_quant = 0.5,
  vel_fix = NULL,
  time_scale = "hours"
)
}
\arguments{
\item{pts}{A prediction or simulation data set produced by \code{\link[crawlUtils]{cu_crw_predict}} or
\code{\link[crawlUtils]{cu_crw_sample}}}

\item{ess}{An effective sample size object. See \verb{\link[crawlUtils]\{cu_crw_ess\}}.}

\item{vel_quant}{A quantile value for the observed average velocities, e.g., 0.5 would imply use of the median
observed average velocity between time points}

\item{vel_fix}{A fixed value for the average velocity used}

\item{time_scale}{The time scale used, e.g., movement per \code{"hours"}.}
}
\description{
Compute velocity based kernel density estimate bandwidth
}
