% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_funcs.R
\name{download_osm}
\alias{download_osm}
\title{Download Open Street Map Coastline Shapefile}
\usage{
download_osm(force = FALSE)
}
\arguments{
\item{force}{Logical. If data has previously been downloaded, it will force a new
download and update of the OSM shapefile.}
}
\description{
Downloads OSM data for plotting telemetry data and use distributions.
}
\references{
https://osmdata.openstreetmap.de/data/land-polygons.html
}
\author{
Josh M. London and Devin S. Johnson
}
