% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_funcs.R
\name{hex_size}
\alias{hex_size}
\title{Calculate cellsize value for hexagon grid}
\usage{
hex_size(area = NULL, radius = NULL, sep = NULL)
}
\arguments{
\item{area}{A value (m^2) for the resulting area of a full hexagon cell}

\item{radius}{The value for the distance (m) from the centroids to the edge of full hexagon cells.}

\item{sep}{The distance (m) between centoids of the hexagon grid.}
}
\description{
Calculates the appropriate \code{cellsize} argument for making
a hexagon grid with \code{\link[sf]{st_make_grid}}.
}
\references{
See \url{https://github.com/r-spatial/sf/issues/1505}
}
\author{
Devin S. Johnson
}
