% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl_utils_funcs.R
\name{location_rate}
\alias{location_rate}
\title{Location rate statistics}
\usage{
location_rate(x, time_name, time_unit = "day", stat = mean, ...)
}
\arguments{
\item{x}{data set containing time of lcoations.}

\item{time_name}{Character name of the POSIX time column when locations are
observed.}

\item{time_unit}{Time unit of the location summary. Defaults to \code{"day"}.}

\item{stat}{Function used to summarize location times. Defaults to \code{mean}.}

\item{...}{Additional arguments passed to \code{stat} function.}
}
\description{
Calculate location rate statistics such as mean location rate per
day or maximum number of locations per day for a telemetry
}
\author{
Devin S. Johnson
}
