% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_funcs.R
\name{st_bbox_list}
\alias{st_bbox_list}
\title{\code{sf::st_bbox} for a list of \code{sf} or \code{sfc} objects.}
\usage{
st_bbox_list(x, union = TRUE, as_sfc = FALSE)
}
\arguments{
\item{x}{A list of \code{sf} or \code{sfc} objects.}

\item{union}{Logical. Should the bounding box of the union be returned instead of
a list of bounding boxes.}

\item{as_sfc}{Logical. Should the bounding box (boxes) be returned as \code{sfc} objects.}
}
\description{
\code{sf::st_bbox} for a list of \code{sf} or \code{sfc} objects.
}
\author{
Devin S. Johnson
}
