% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_funcs.R
\name{st_expand}
\alias{st_expand}
\title{Expand Spatial Bounding Box}
\usage{
st_expand(bbox, ef)
}
\arguments{
\item{bbox}{An \code{sf} bounding box. See \code{\link[sf:st_bbox]{sf::st_bbox}}.}

\item{ef}{Expansion factor, must be positive}
}
\description{
Expand an \code{sf} bounding box by an expansion factor
}
\author{
Josh M. London
}
