% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_funcs.R
\name{st_not_within}
\alias{st_not_within}
\title{Predicate function for st_filter}
\usage{
st_not_within(x, y, sparse = TRUE, prepared = TRUE, ...)
}
\arguments{
\item{x}{object of class sf, sfc or sfg}

\item{y}{object of class sf, sfc or sfg; if missing, x is used}

\item{sparse}{ogical; should a sparse index list be returned (TRUE) or a dense logical matrix? See \link[sf:st_within]{sf::st_within}.}

\item{prepared}{ogical; prepare geometry for x, before looping over y? See \link[sf:st_within]{sf::st_within}.}

\item{...}{passed on to s2_options}
}
\description{
Predicate function to use with \code{st_filter} such that
such that elements of one spatial object are selected if
they are not contained at all in the other. See \code{\link[sf:st_within]{sf::st_within}}
}
