% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_funcs.R
\name{st_union_list}
\alias{st_union_list}
\title{\code{sf::st_union} for a list of \code{sf} or \code{sfc} objects.}
\usage{
st_union_list(x)
}
\arguments{
\item{x}{A list of \code{sf} or \code{sfc} objects.}
}
\description{
\code{sf::st_union} for a list of \code{sf} or \code{sfc} objects.
}
\author{
Devin S. Johnson
}
