% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_funcs.R
\name{cu_crw_ess}
\alias{cu_crw_ess}
\title{Calculate Effective Sample Size for a Set of CRW locations}
\usage{
cu_crw_ess(fit, aug = NULL)
}
\arguments{
\item{fit}{A \code{crwFit} object (See \code{\link[crawl]{crwMLE}}).}

\item{aug}{Either a \code{\link[crawl]{crwPredict}} or \code{\link[crawl]{crwPostIS}} objects
from which the extra \code{predTime} location times will be used in the calculation.
The \code{\link[crawl]{crw_as_sf}} transformed versions of these objects will also work.}
}
\description{
Estimates the number of independent locations in a CRW data set
using the method of Acosta and Vallejos (2018) \code{AV18}.
}
\details{
The AV18 method was designed for spatial regression analysis, but
the derivations only use a general correlation matrix. Therefore, the time-series
correlation matrix of the CRW (IOU) process was substituted. However, there is one
change. The CRW (IOU) model is not stationary, so the correlation matrix of the observations
conditioned on the first observations is used. The resulting sample size is then
incremented by 1 to account for the first observation. The first observation
can be regarded as a single independent observation from the animal UD, the AV18
calculation then adds the number of additional 'independent' observations given the
realization of the first one.
}
\references{
Acosta, J., & Vallejos, R. (2018). Effective sample size for
spatial regression models. Electronic Journal of Statistics, 12:3147-3180.
}
\author{
Devin S. Johnson
}
