% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_osm_download.R
\name{cu_download_osm}
\alias{cu_download_osm}
\title{Download Open Street Map Coastline Shapefile from Daylight Map Distribution}
\usage{
cu_download_osm(force = FALSE, save_as_sf = TRUE)
}
\arguments{
\item{force}{Logical. If data has previously been downloaded, it will force a new
download and update of the OSM shapefile.}

\item{save_as_sf}{Logical. Should the world poolygon data be saved as
an \code{sf} polygon data frame? This can speed subsetting in the
\code{\link[crawlUtils]{cu_osm_coast}} function.}
}
\description{
Downloads OSM data for plotting telemetry data and use distributions.
}
\references{
https://osmdata.openstreetmap.de/data/land-polygons.html
}
\author{
Josh M. London and Devin S. Johnson
}
