% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_empirical_ud.R
\name{cu_empirical_ud}
\alias{cu_empirical_ud}
\title{Calculate an empirical utilization distribution}
\usage{
cu_empirical_ud(pts, grid, average = TRUE)
}
\arguments{
\item{pts}{an sf point object}

\item{grid}{a  \code{SpatRaster} on which to evaluate the UD.}

\item{average}{should an average UD be created from 1 or more pts layers}
}
\value{
grid with an additional column, \emph{npts} or \emph{mean_pts}
}
\description{
Calculate an empirical utilization distribution
}
