% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_gmm_ud_sample.R
\name{cu_gmm_ud_sample}
\alias{cu_gmm_ud_sample}
\title{Kernel density estimation for a posterior sample list}
\usage{
cu_gmm_ud_sample(smp_list, grid, ess = NULL, ...)
}
\arguments{
\item{smp_list}{A list of posterior track samples from \code{\link[crawl]{crwPostIS}} after
transformation via \code{\link[crawl]{crw_as_sf}}.}

\item{grid}{An \code{\link[sf]{sf}} data frame containing the desired grid location for UD estimation.}

\item{ess}{Effective sample size.}

\item{...}{Additional arguments passed to \code{\link[mclust]{densityMclust}}.}
}
\description{
Kernel density estimation for a posterior sample list
}
\author{
Devin S. Johnson
}
