% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_kde_ud.R
\name{cu_kde_ud}
\alias{cu_kde_ud}
\title{Kernel density estimator}
\usage{
cu_kde_ud(
  pts,
  grid,
  ess = NULL,
  use_w = TRUE,
  norm = TRUE,
  bw = NULL,
  bw_subset = TRUE,
  type = "original",
  ...
)
}
\arguments{
\item{pts}{A \code{\link[crawl]{crwPredict}} or \code{\link[crawl]{crwPostIS}} object, or
their 'sf' versions (See \link[crawl]{crw_as_sf}).}

\item{grid}{An \code{\link[sf]{sf}} data frame containing the desired grid location for UD estimation.}

\item{ess}{Effective sample size object or a \code{crwFit} object. See \verb{link[crawlUtils]\{cu_crw_ess\}}.}

\item{use_w}{Use weights from the \code{ess} object for a weighted KDE.}

\item{norm}{Logical. Should each individual kernel be normalized to
sum-to-one over the locations in \code{grid}. Defaults to \code{kern = TRUE}}

\item{bw}{Kernel bandwidth (standard deviation of Gaussian kernel). Defaults to the
default plugin bandwidth \code{bandwidth.nrd} in the \code{MASS} package.}

\item{bw_subset}{A vector of values indicating which \code{pts} should be used for calculating \code{B} if left unspecified.}

\item{type}{Form of the return type. Can be \code{"original"} to have the UD returned in the same form as the \code{grid}
argument. Or set \code{type="vector"} to return only a vector of UD values.}

\item{...}{additional arguments passed to \code{\link[crawlUtils]{cu_vel_B}}}
}
\description{
Calculates a Gaussian KDE for \code{crawl} simulations and predictions.
}
\author{
Devin S. Johnson
}
