% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_funcs.R
\name{cu_kde_ud}
\alias{cu_kde_ud}
\title{Calculate Gaussian Kernel Utilization Distribution}
\usage{
cu_kde_ud(x, barrier = NULL, norm, ...)
}
\arguments{
\item{x}{an sf or sfc object. The KDE grid will be estimated from the bounding box.}

\item{barrier}{An sf polygon object that defines an area where use is excluded.}

\item{norm}{Logical. Should each individual kernel be normalized to account for barriers
before addition to the total KDE. Defaults to \code{TRUE} if a barrier is specified.}

\item{...}{Any arguments passed to \code{sf::st_make_grid} to create the KDE prediction grid. See \link[sf:st_make_grid]{sf::st_make_grid} for
description of additional arguments to make the KDE grid.}
}
\description{
Calculates a kernel density estimate using telemetry locations
over a grid accounting for barriers by normalizing each kernel before addition
to the overall sum. The final estimate will always be normalized to sum to one
over the estimate grid.
}
\author{
Devin S. Johnson
}
