% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_funcs.R
\name{cu_osm_coast}
\alias{cu_osm_coast}
\title{Get A Coastline \code{sf} Polygon Object For Plotting and Mapping}
\usage{
cu_osm_coast(x, keep = 0.2)
}
\arguments{
\item{x}{An \code{sf} spatial object. The coastline will be cropped to the bounding box.}

\item{keep}{The amount of data retained after simplification with \code{\link[rmapshaper:ms_simplify]{rmapshaper::ms_simplify}}}
}
\description{
Uses downloaded OSM data for constructing an \code{sf} polygon coastline data object.
Prior to using this function you must run \code{crawlUtils::download_osm()}.
See \code{\link[crawlUtils:cu_download_osm]{crawlUtils::download_osm}}.
}
\author{
Josh M. London and Devin S. Johnson
}
