% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_indiv_files_wc.R
\name{cu_read_wc_dirs}
\alias{cu_read_wc_dirs}
\title{Read individual telemetry data from Wildlife Computers portal directories}
\usage{
cu_read_wc_dirs(x)
}
\arguments{
\item{x}{Directory containing the individual telemetry data directories.}
}
\description{
Read and combine data downloaded from Wildlife Computers portal
into individual directories.
}
\author{
Devin S. Johnson
}
