% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_read_wc_dirs.R
\name{cu_read_wc_dirs}
\alias{cu_read_wc_dirs}
\title{Read individual telemetry data from Wildlife Computers portal directories}
\usage{
cu_read_wc_dirs(x, remove_duplicates = TRUE)
}
\arguments{
\item{x}{Directory containing the individual telemetry data directories.}

\item{remove_duplicates}{Locgical. Should observations with duplicated times be removed? The observation
with the highest quality will be retained.}
}
\description{
Read and combine data downloaded from Wildlife Computers portal
into individual directories.
}
\author{
Devin S. Johnson, Josh M. London
}
