% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu_ud_contour.R
\name{cu_ud_contour}
\alias{cu_ud_contour}
\title{Approximate contours for a UD}
\usage{
cu_ud_contour(ud, prob = seq(0.9, 0.1, -0.1), trim_pixel = TRUE)
}
\arguments{
\item{ud}{A \code{SpatRaster} created by \verb{\link[crawlUtils]\{cu_sf_kde\}}}

\item{prob}{a vector of probabilities for the contours}

\item{trim_pixel}{Logical. Should contour areas less than a pixel size be trimmed? defaults to \code{TRUE}.}
}
\description{
Approximate contours for a UD
}
\author{
Devin S. Johnson
}
